requirements.txt
Django==2.1.4
django-nose==1.4.6
djangorestframework==3.9.0
nose==1.3.7
psycopg2==2.7.6.1
pytz==2018.7
